services/prompt-enricher/requirements.txt
pyspark==4.0.0
google-genai
dotenv
python-logstash
python-json-logger
beautifulsoup4